Found a bunch of "int time" that were changed to "double time"

+++ JBotSim/src/simulation/environment/TwoNestCenterDeployPreysEnvironment.java
@@ -64,7 +64,7 @@ public class TwoNestCenterDeployPreysEnvironment extends
 	}
 
 	@Override
-	public void update(int time) {
+	public void update(double time) {
 		super.update(time);
 		reDeployPreys();
 	}
